Delay analyzing image data until needed
Before, when a "frame" was created, the image data from the canvas was
immediately analyzed and converted into GIF-formatted bytes. But we
don't want to immediately convert the image data to bytes because the
user may still be drawing their image.

Also made a Frame class rather than having a function that just returns
the frame objects (this just seems a bit more ES6 and better
captures/documents the intended structure of frame objects).