Beef up getColorData
There are several parameters that we need to put into a GIF that
ultimately depend on the size of the color table, but that were being
calculated repeatedly across different functions. This was confusing as
well as redundant.

Updated getColorData to return an object containing not only the color
table but also the minCodeSize in bits needed to represent all of those
colors, and the "encoded" color table size (i.e., the integer between 0
and 7 inclusive that represents the table size within the GIF).

Also did some general cleanup in getCodeReader, added some comments.