Bug fix for 1-2 color images
GIFs produced with only 1 or 2 colors weren't being parsed by Chromium
(we ended up with a "blank" GIF that was all transparent).

We were outputting a minimum code size of 1 with 2 colors, and 0 with 1.
However, when GIMP outputs a GIF with only one or two colors, the
minimum code size comes out as 2.

So, making sure that the minimum code size is at least 2. When the
minimum code size is 2+, the GIFs get parsed as intended.

With this change, we also have to make sure getCodeReader returns codes
of the expected sizes. So if we have only have 2 colors, tell
getCodeReader we actually have 4.