Stop sharing "tracking canvas" between components
By having a "tracking canvas" in the state of GifEditor (as an attribute
of a frame),  and then modifying it in the DrawingCanvas component, we
were violating the rule against directly modifying state and props,
which per the React docs can lead to subtle bugs. Avoiding this by
having DrawCanvas pass updated image information to the drawingUpdated
method in its props (so something like making the canvas a "controlled
component" - though we're only updating everything when the user
finishes a line, to avoid performance issues).

This is also probably clearer and more explicit, specifically updating
the state to include the new, updated image with setState, rather than
having the canvas magically change and then having to "trick" React into
updating.