Move some attributes on DrawCanvas out of state
A couple of variables - lastPoint and the mouse down status - were in
DrawCanvas' state, but didn't need to be, leading to unnecessary updates
whenever we changed them. Now that we're avoiding those updates, we
don't need to check the mouse status before proceeding in
componentDidUpdate.

(Renamed mouseDown attribute to mouseIsDown due to name collision with
the mouseDown method).