Make brush size adjustment work better
The brush size specified by the user now specifies the approximate
diameter of the resulting brush instead of the radius (for finer grained
control). Some tweaks were made to the brush generating calculations to
make the diameter closer to what is specified. Also a small tweak was
made to the DrawCanvas component to prevent us from drawing the brush on
fractional pixels, which has weird, distorted results.